Added WithEnvironment overload supporting Keyvault refs - Added tests - Moved KV tests to a new file
